{Note -> Line.Note}, {Roll.RollNote -> Roll.Note}
